backend/pyproject.toml
[tool.poetry]
name = "calendar-backend"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
pydantic = "^1.10.2"
requests = "^2.28.1"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.3"
asyncpg = "^0.26.0"
SQLAlchemy = "^1.4.41"
databases = "^0.6.1"
gunicorn = "^20.1.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^10.1.4"}
greenlet = "1.1.3"
python-dateutil = "2.8.2"
bitarray = "2.6.0"


[tool.poetry.dev-dependencies]
black = "^22.8.0"
pytest = "^7.1.3"
ipython = "^8.5.0"
pytest-factoryboy = "2.5.0"
pytest-cov = "4.0.0"
pytest-forked = "^1.4.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.cruft]
skip = [".env"]
